+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Adds `new_id` function
 - Deprecated `next_id` in favor of `new_id`
 - Adds `identity` and `has_identity` functions
+- Adds `singleton` and `events` functions
 
 ## 1.1.2 (July 06, 2017)
 - Fix typescript declaration file to include default types